+++ examples/Effects/Effects.ino
@@ -9,12 +9,16 @@ void setup() {
 	Serial.begin(9600);
 	while (!Serial);
 
+	//init AxeSystem
+	Axe.init(Serial1);
+
 	//register a callback so we know when preset details are available
 	Axe.registerPresetChangeCallback(onPresetChange);
 
 	//request current preset
 	Axe.requestPresetDetails();
 
+	//ask for refresh call
 	Axe.enableRefresh(5000);
 
 }
