+++ src/AxeSystem_Commands.cpp
@@ -1,19 +1,67 @@
 #include "AxeSystem.h"
 #include "AxeEffect.h"
 
+		const byte REQUEST_EFFECT_BYPASSED_COMMAND_9_BYTES[9] = {
+				SYSEX_MANUFACTURER_BYTE1,  		SYSEX_MANUFACTURER_BYTE2,
+				SYSEX_MANUFACTURER_BYTE3,  		SYSEX_AXE_VERSION,
+				SYSEX_REQUEST_EFFECT_BYPASS, 	SYSEX_QUERY_BYTE,
+				SYSEX_QUERY_BYTE,          		SYSEX_QUERY_BYTE,
+				SYSEX_CHECKSUM_PLACEHOLDER};
+
+		const byte REQUEST_SCENE_NAME_COMMAND_7_BYTES[7] = {
+				SYSEX_MANUFACTURER_BYTE1,  SYSEX_MANUFACTURER_BYTE2,
+				SYSEX_MANUFACTURER_BYTE3,  SYSEX_AXE_VERSION,
+				SYSEX_REQUEST_SCENE_INFO,  SYSEX_QUERY_BYTE,
+				SYSEX_CHECKSUM_PLACEHOLDER};
+
+		const byte REQUEST_SCENE_NUMBER_COMMAND_7_BYTES[7] = {
+				SYSEX_MANUFACTURER_BYTE1,   SYSEX_MANUFACTURER_BYTE2,
+				SYSEX_MANUFACTURER_BYTE3,   SYSEX_AXE_VERSION,
+				SYSEX_REQUEST_SCENE_NUMBER, SYSEX_QUERY_BYTE,
+				SYSEX_CHECKSUM_PLACEHOLDER};
+
+		const byte REQUEST_TEMPO_COMMAND_8_BYTES[8] = {
+				SYSEX_MANUFACTURER_BYTE1, SYSEX_MANUFACTURER_BYTE2,
+				SYSEX_MANUFACTURER_BYTE3, SYSEX_AXE_VERSION,
+				SYSEX_REQUEST_TEMPO,      SYSEX_QUERY_BYTE,
+				SYSEX_QUERY_BYTE,         SYSEX_CHECKSUM_PLACEHOLDER};
+
+		const byte TAP_TEMPO_PULSE_COMMAND_6_BYTES[6] = {
+				SYSEX_MANUFACTURER_BYTE1, SYSEX_MANUFACTURER_BYTE2,
+				SYSEX_MANUFACTURER_BYTE3, SYSEX_AXE_VERSION,
+				SYSEX_TAP_TEMPO_PULSE,    SYSEX_CHECKSUM_PLACEHOLDER};
+
+		const byte ENABLE_TUNER_COMMAND_7_BYTES[7] = {
+				SYSEX_MANUFACTURER_BYTE1,	SYSEX_MANUFACTURER_BYTE2,
+				SYSEX_MANUFACTURER_BYTE3,	SYSEX_AXE_VERSION,
+				SYSEX_TUNER,							SYSEX_TUNER_ON,
+				SYSEX_CHECKSUM_PLACEHOLDER};
+
+		const byte DISABLE_TUNER_COMMAND_7_BYTES[7] = {
+				SYSEX_MANUFACTURER_BYTE1,	SYSEX_MANUFACTURER_BYTE2,
+				SYSEX_MANUFACTURER_BYTE3,	SYSEX_AXE_VERSION,
+				SYSEX_TUNER,							SYSEX_TUNER_OFF,
+				SYSEX_CHECKSUM_PLACEHOLDER};
+
+		const byte REQUEST_LOOPER_STATUS_COMMAND_7_BYTES[7] = {
+				SYSEX_MANUFACTURER_BYTE1,  		SYSEX_MANUFACTURER_BYTE2,
+				SYSEX_MANUFACTURER_BYTE3,  		SYSEX_AXE_VERSION,
+				SYSEX_REQUEST_LOOPER_STATUS,  SYSEX_QUERY_BYTE,
+				SYSEX_CHECKSUM_PLACEHOLDER};
+
 void AxeSystem::requestFirmwareVersion() {
-	sendSysEx(6, (byte*) REQUEST_FIRMWARE_VERSION_COMMAND_6_BYTES);
+	sendCommand(SYSEX_REQUEST_FIRMWARE);
 }
 
 void AxeSystem::requestPresetName(int preset) {
-  if (preset == -1) {
-    sendSysEx(8, (byte*) REQUEST_PRESET_NAME_COMMAND_8_BYTES);
-  } else {
-    byte command[8];
-    memcpy(command, REQUEST_PRESET_NAME_COMMAND_8_BYTES, 8);
-    intToMidiBytes(preset, &command[5], &command[6]);
-    sendSysEx(8, (byte*) command);
+
+	byte data[2] = {SYSEX_QUERY_BYTE};
+
+  if (preset != -1) {
+    intToMidiBytes(preset, &data[0], &data[1]);
   }
+
+	sendCommand(SYSEX_REQUEST_PRESET_INFO, data, 2);
 }
 
 void AxeSystem::requestSceneName() {
@@ -24,8 +72,9 @@ void AxeSystem::requestSceneNumber() {
 	sendSysEx(7, (byte*) REQUEST_SCENE_NUMBER_COMMAND_7_BYTES);
 }
 
+
 void AxeSystem::requestEffectDetails() {
-	sendSysEx(6, (byte*) REQUEST_EFFECT_DUMP_6_BYTES);
+	sendCommand(SYSEX_EFFECT_DUMP);
 }
 
 void AxeSystem::sendSceneChange(byte scene) {
