+++ src/AxeSystem_Midi.cpp
@@ -1,0 +1,126 @@
+#include <AxeSystem.h>
+
+#define SER_AVLB _serial->available()
+#define SER_READ _serial->read()
+#define SER_SEND _serial->write
+
+void AxeSystem::readMidi() {
+
+	if (SER_AVLB) {
+		
+		byte data = SER_READ;
+
+		if (_readingSysex) {
+			_sysexBuffer[_sysexCount++] = data;
+			if (data == SYSEX_END) {
+				_readingSysex = false;
+				if (isAxeSysEx(_sysexBuffer, _sysexCount)) {
+					onSystemExclusive(_sysexBuffer, _sysexCount);
+				} else {
+					_sysexCount = 0;
+				}
+			}
+
+		} else {
+
+			switch (data) {
+
+				case ControlChange: {
+					while (SER_AVLB < 2);
+					if (BANK_CHANGE_CC == SER_READ) {
+						_bank = SER_READ;
+					}
+					break;
+				}
+
+				case ProgramChange: {
+					while (SER_AVLB < 1);
+					byte patch = SER_READ;
+					onPresetChange(_bank * 128 + patch);
+					break;
+				}
+
+				case SystemExclusive: {
+					_readingSysex = true;
+					_sysexCount = 0;
+					_sysexBuffer[_sysexCount++] = data;
+					break;
+				}
+
+			}
+
+		}
+
+	}
+
+}
+
+void AxeSystem::sendPresetChange(const unsigned number) {
+
+	SER_SEND(ControlChange);
+	SER_SEND(BANK_CHANGE_CC);
+	SER_SEND(number / 128);
+	SER_SEND(ProgramChange);
+	SER_SEND(number % 128);
+
+	setChanging();
+	callPresetChangingCallback(number);
+
+}
+
+void AxeSystem::sendCommand(const byte command) {
+	sendCommand(command, nullptr, 0);
+}
+
+void AxeSystem::sendCommand(const byte command, const byte *data, const byte paramCount) {
+	
+	byte length = 0;
+	byte sysex[paramCount + 8];
+	
+	//header
+	sysex[length++] = SystemExclusive;
+	sysex[length++] = SYSEX_MANUFACTURER_BYTE1;
+	sysex[length++] = SYSEX_MANUFACTURER_BYTE2;
+	sysex[length++] = SYSEX_MANUFACTURER_BYTE3;
+	sysex[length++] = SYSEX_AXE_VERSION;
+	sysex[length++] = command;
+
+	//optional data
+	for (byte j=0; j<paramCount; j++) {
+		sysex[length++] = data[j];
+	}
+
+	//checksum
+	byte sum = 0xF0;
+	for (int i=1; i<length; ++i) {
+		sum ^= sysex[i];
+	}
+
+	//footer
+	sysex[length++] = (sum & 0x7F);
+	sysex[length++] = SYSEX_END;
+
+	//punch it!
+	for (byte i=0; i<length; i++) {
+		SER_SEND(sysex[i]);
+	}
+
+	#ifdef AXE_DEBUG_SYSEX
+	char buf[6];
+	Serial.print("SYSEX OUT: ");
+	for (byte i=0; i<length; i++) {
+		snprintf(buf, 6, "0x%02X ", sysex[i]);
+		Serial.print(buf);
+	}
+	Serial.println();
+	#endif
+
+}
+
+bool AxeSystem::isAxeSysEx(const byte *sysex, const unsigned length) {
+	return 
+		length > 4 && 
+		sysex[1] == SYSEX_MANUFACTURER_BYTE1 && 
+		sysex[2] == SYSEX_MANUFACTURER_BYTE2 && 
+		sysex[3] == SYSEX_MANUFACTURER_BYTE3;
+}
